Update the sdk to support the latest changes

+++ Openfort.SDK/Model/PlayerCreateAccountRequest.cs
@@ -2,7 +2,7 @@
 {
 	public class PlayerCreateAccountRequest: CreatePlayerAccountRequest
     {
-		public PlayerCreateAccountRequest(string playerId, int chainId = default(int), string externalOwnerAddress = null, DataAccountTypes? accountType = null, string tokenContract = null, double tokenId = default(double)): base(chainId, externalOwnerAddress, accountType, tokenContract, tokenId)
+		public PlayerCreateAccountRequest(string playerId, int chainId = default(int), string externalOwnerAddress = null, DataAccountTypes? accountType = null, string tokenContract = null, long tokenId = default(long)): base(chainId, externalOwnerAddress, accountType, tokenContract, tokenId)
 		{
 			PlayerId = playerId;
 		}
